Added reconciliation and winning bid assertion
- Added retry logic for better handling of bid races
- Improved tests

+++ src/main/java/com/jpswcons/auctioneer/data/entities/Bid.java
@@ -24,7 +24,7 @@ public class Bid {
     private long id;
     private long auctionId;
     private long bidderId;
-    private long amount;
+    private int amount;
     @CreationTimestamp
     private LocalDateTime createdTime;
     @UpdateTimestamp
